One platform header; each target keeps only its clock (S1; F-17, F-18, N-39)
There was a platform.h per target, plus copies for the unit tier (two)
and the tools (two), and they had drifted.  The wasm copy hand-rolled
the bit scans the headless one got from the compiler.  Both carried a
mutex, a thread id, drive-activity enums, a tick counter and a screen
hook that nothing used, and the unit tier defined platform_bsr32/ntz32
seven more times across six files.

- src/platform/platform.h is shared: byte swaps, MIN/MAX, the bit scans
  as __builtin, and the audio and callstack prototypes.  It includes
  platform_clock.h from the target's include path: CLOCK_MONOTONIC on
  headless, emscripten_get_now on wasm, zero in the unit tier, and a
  controllable fake in the scheduler suite.
- Both builds put src/platform ahead of the target directory.  The
  tools force-include the shared header with the headless clock; their
  private copies are gone.
- Headless implements the audio no-ops and the callstack as functions,
  and the glibc backtrace moves there from em_main.c's never-compiled
  native branch.
- The unit tier's platform.h wraps the shared one, and the private
  bit-scan definitions (stub_platform.c and five test files) are deleted.
- memory_debug_read_phys compared a cast int page with the unsigned page
  count; it compares unsigned values now (a sign-compare warning D3
  added).

+++ src/core/memory/memory.c
@@ -735,7 +735,7 @@ uint32_t memory_debug_read_phys(uint32_t phys, unsigned size, bool *ok) {
     for (unsigned i = 0; i < size; i++) { // byte-wise: an access may straddle pages
         uint32_t a = phys + i;
         uint32_t page = a >> PAGE_SHIFT;
-        if ((int)page >= g_page_count)
+        if (page >= g_page_count)
             return 0;
         page_entry_t *pe = &g_page_table[page];
         uint8_t b;
